+++ wrappers/interface/ITextTrackWrapper.h
@@ -129,7 +129,7 @@ public:
      *
      * @retval the error code
      */
-    virtual std::uint32_t sendSessionData(std::uint32_t sessionId, DataType type, std::int32_t displayOffsetMs,
+    virtual std::uint32_t sendSessionData(std::uint32_t sessionId, DataType type, std::int64_t displayOffsetMs,
                                           const std::string &data) const = 0;
 
     /**
